AddressProvider for United States of America

Object to create addresses for a locale. Makes use of PersonProvider for creating street names.

Value

Returns an AddressProvider object.

Details

When there is no PersonProvider for this locale, we default back to en_US.

See also

Super classes

charlatan::BareProvider -> charlatan::BaseProvider -> charlatan::AddressProvider -> AddressProvider_en_US

Methods

Public methods

AddressProvider_en_US$address()

AddressProvider_en_US$civ_address()

AddressProvider_en_US$mil_address()

AddressProvider_en_US$city()

AddressProvider_en_US$street_name()

AddressProvider_en_US$street_address()

AddressProvider_en_US$postcode()

AddressProvider_en_US$building_number()

AddressProvider_en_US$state()

AddressProvider_en_US$clone()

Inherited methods

charlatan::BareProvider$bothify()

charlatan::BareProvider$lexify()

charlatan::BareProvider$numerify()

charlatan::BareProvider$random_digit()

charlatan::BareProvider$random_digit_not_zero()

charlatan::BareProvider$random_digit_not_zero_or_empty()

charlatan::BareProvider$random_digit_or_empty()

charlatan::BareProvider$random_element()

charlatan::BareProvider$random_element_prob()

charlatan::BareProvider$random_int()

charlatan::BareProvider$random_letter()

charlatan::BareProvider$randomize_nb_elements()

charlatan::BaseProvider$allowed_locales()

charlatan::BaseProvider$check_locale()

charlatan::BaseProvider$print()

charlatan::AddressProvider$init_person_provider()

charlatan::AddressProvider$initialize()

Method address()

address

Usage

AddressProvider_en_US$address()

Method civ_address()

civilian address, the type of address you would expect. Not to be confused with Military address which is also available for this locale.

Usage

AddressProvider_en_US$civ_address()

Method mil_address()

Military address

Usage

AddressProvider_en_US$mil_address()

Method city()

city

Usage

AddressProvider_en_US$city()

Method street_name()

street name

Usage

AddressProvider_en_US$street_name()

Method street_address()

street address

Usage

AddressProvider_en_US$street_address()

Method postcode()

postal code

Usage

AddressProvider_en_US$postcode()

Method building_number()

building number

Usage

AddressProvider_en_US$building_number()

Method state()

state

Usage

AddressProvider_en_US$state()

Method clone()

The objects of this class are cloneable with this method.

Usage

AddressProvider_en_US$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

(z <- AddressProvider_en_US$new())
#> < AddressProvider >
#> locale: en_US
z$locale
#> [1] "en_US"
z$postcode()
#> [1] "28584"
z$street_name()
#> [1] "Vincenzo Shoal"
z$address()
#> [1] "8122 Kassulke Causeway Suite 643\nNorth Lolitaburgh, MI 69750-3154"
z$city()
#> [1] "West Zoe"
z$mil_address()
#> [1] "PSC 1842, Box 1493\nAPO AA 89828"
z$civ_address()
#> [1] "06483 Durrell Forest\nSporerton, CO 90491-8806"