Internet provider

methods for internet related data, like email addresses, usernames, and websites.

See also

Super classes

charlatan::BareProvider -> charlatan::BaseProvider -> charlatan::InternetProvider -> InternetProvider_de_DE

Methods

Public methods

InternetProvider_de_DE$clone()

Inherited methods

charlatan::BareProvider$bothify()

charlatan::BareProvider$lexify()

charlatan::BareProvider$numerify()

charlatan::BareProvider$random_digit()

charlatan::BareProvider$random_digit_not_zero()

charlatan::BareProvider$random_digit_not_zero_or_empty()

charlatan::BareProvider$random_digit_or_empty()

charlatan::BareProvider$random_element()

charlatan::BareProvider$random_element_prob()

charlatan::BareProvider$random_int()

charlatan::BareProvider$random_letter()

charlatan::BareProvider$randomize_nb_elements()

charlatan::BaseProvider$allowed_locales()

charlatan::BaseProvider$check_locale()

charlatan::BaseProvider$print()

charlatan::InternetProvider$ascii_company_email()

charlatan::InternetProvider$ascii_email()

charlatan::InternetProvider$ascii_free_email()

charlatan::InternetProvider$ascii_safe_email()

charlatan::InternetProvider$company_email()

charlatan::InternetProvider$domain_name()

charlatan::InternetProvider$domain_word()

charlatan::InternetProvider$email()

charlatan::InternetProvider$free_email()

charlatan::InternetProvider$free_email_domain()

charlatan::InternetProvider$image_url()

charlatan::InternetProvider$initialize()

charlatan::InternetProvider$ipv4()

charlatan::InternetProvider$ipv6()

charlatan::InternetProvider$mac_address()

charlatan::InternetProvider$safe_email()

charlatan::InternetProvider$slug()

charlatan::InternetProvider$tld()

charlatan::InternetProvider$to_ascii()

charlatan::InternetProvider$uri()

charlatan::InternetProvider$uri_extension()

charlatan::InternetProvider$uri_page()

charlatan::InternetProvider$uri_path()

charlatan::InternetProvider$url()

charlatan::InternetProvider$user_name()

Method clone()

The objects of this class are cloneable with this method.

Usage

InternetProvider_de_DE$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- InternetProvider_de_DE$new()
#> Warning:  LoremProvider does not have locale de_DE, defaulting to en_US locale.
x$email()
#> [1] "ibrahim37@yahoo.de"
x$free_email()
#> [1] "dblumel@hotmail.de"
x$mac_address()
#> [1] "9c:cb:97:b3:8b:b5:de"
x$company_email()
#> [1] "plath.niko@faust.de"