Lorem provider Arabic

Methods for Lorem Ipsum generation. Lorem Ipsum is a placeholder text commonly used to demonstrate the visual form of a document or a typeface without relying on meaningful content.

See also

Super classes

charlatan::BareProvider -> charlatan::BaseProvider -> charlatan::LoremProvider -> LoremProvider_ar_AA

Methods

Public methods

LoremProvider_ar_AA$clone()

Inherited methods

charlatan::BareProvider$bothify()

charlatan::BareProvider$lexify()

charlatan::BareProvider$numerify()

charlatan::BareProvider$random_digit()

charlatan::BareProvider$random_digit_not_zero()

charlatan::BareProvider$random_digit_not_zero_or_empty()

charlatan::BareProvider$random_digit_or_empty()

charlatan::BareProvider$random_element()

charlatan::BareProvider$random_element_prob()

charlatan::BareProvider$random_int()

charlatan::BareProvider$random_letter()

charlatan::BareProvider$randomize_nb_elements()

charlatan::BaseProvider$allowed_locales()

charlatan::BaseProvider$check_locale()

charlatan::BaseProvider$print()

charlatan::LoremProvider$initialize()

charlatan::LoremProvider$paragraph()

charlatan::LoremProvider$paragraphs()

charlatan::LoremProvider$sentence()

charlatan::LoremProvider$sentences()

charlatan::LoremProvider$text()

charlatan::LoremProvider$word()

charlatan::LoremProvider$words()

Method clone()

The objects of this class are cloneable with this method.

Usage

LoremProvider_ar_AA$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- LoremProvider_ar_AA$new()
x$word()
#> [1] "الصعداء"
x$words(3)
#> [1] "ولكسمبورغ" "بفرض"      "حالية"    
x$words(6)
#> [1] "الحكم"   "ذات"     "لبلجيكا" "لان"     "عرفها"   "قِبل"    
x$sentence()
#> [1] "بدارت الإمتعاض بمما المتاخمة على وانهاء."
x$paragraph()
#> [1] "العالمية الفرنسي حالية تمهيد تُصب المعاهدات وبعض نتيجة. الخاسر به دار و مسارح."
x$paragraphs(3)
#> [1] "بعلى بحث أفريقيا وحرمان. بسبب وبعض اتّجة بالسيطرة."                                                   
#> [2] "بضرب بالمحور للجزر لعملة مسؤولية. السبب اعتداء المارق. الشّعبين وصغار موالية واتّجه."                  
#> [3] "الخطّة مرمى وحتى. إعمار شرسة والفلبين ألمّ. ضمنها أخر فقامت تعديل شموليةً كما. ببعض ليرتفع ببحشد بتحدّي."
cat(x$paragraphs(6), sep = "\n")
#> العالمي وبحلول مشاركة أفريقيا بقسوة بكلّ. غضون اليابانية أدوات.
#> الصينية وبدأت بالرّغم الشرق أملاً ماليزيا المارق. لليابان اعتداء ديسمبر عُقر وايرلندا سقطت عشوائية. غضون التقليدية المتاخمة ماليزيا ماشاء الله بكلّ.
#> لدحر ووصف إتفاقية ثانية لكون كلّ ديسمبر. الحدود قررت تجهيز الأرض ليرتفع.
#> بمباركة مما شعار استبدال. مسرح بانه انه وبداية والعتاد. الصعداء وانهاء بقسوة وإقامة وتنصيب الشرق العظمى.
#> الصين الأوروبية بالجوي مقاطعة. عُقر قائمة لدحر اليميني الصينية بنقطة شمال. الساحة السيطرة الآخر للسيطرة.
#> ولاتّساع حتى وباءت بولندا تطوير مايو الجنود. والحزب زهاء الساحة غرّة ببحشد ارتكبها جهة بشرية.
x$text(19)
#> [1] "التاريخ إتفاقية."
x <- LoremProvider_ar_AA$new(word_connector = " --- ")
x$paragraph(4)
#> [1] "نفس --- ونتج --- اليميني --- جسيمة --- بحيث --- لفشل. --- جديدة --- وقام --- للسيطرة --- وعُرفت. --- ساعة --- ومطالبة --- وعُرفت --- التبرعات --- لغات --- الهجوم --- بأيدي."