Lorem provider

Methods for Lorem Ipsum generation. Lorem Ipsum is a placeholder text commonly used to demonstrate the visual form of a document or a typeface without relying on meaningful content.

See also

Super classes

charlatan::BareProvider -> charlatan::BaseProvider -> charlatan::LoremProvider -> LoremProvider_he_IL

Methods

Public methods

LoremProvider_he_IL$clone()

Inherited methods

charlatan::BareProvider$bothify()

charlatan::BareProvider$lexify()

charlatan::BareProvider$numerify()

charlatan::BareProvider$random_digit()

charlatan::BareProvider$random_digit_not_zero()

charlatan::BareProvider$random_digit_not_zero_or_empty()

charlatan::BareProvider$random_digit_or_empty()

charlatan::BareProvider$random_element()

charlatan::BareProvider$random_element_prob()

charlatan::BareProvider$random_int()

charlatan::BareProvider$random_letter()

charlatan::BareProvider$randomize_nb_elements()

charlatan::BaseProvider$allowed_locales()

charlatan::BaseProvider$check_locale()

charlatan::BaseProvider$print()

charlatan::LoremProvider$initialize()

charlatan::LoremProvider$paragraph()

charlatan::LoremProvider$paragraphs()

charlatan::LoremProvider$sentence()

charlatan::LoremProvider$sentences()

charlatan::LoremProvider$text()

charlatan::LoremProvider$word()

charlatan::LoremProvider$words()

Method clone()

The objects of this class are cloneable with this method.

Usage

LoremProvider_he_IL$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- LoremProvider_he_IL$new()
x$word()
#> [1] "דולור"
x$words(3)
#> [1] "וואל"       "קונדימנטום" "סטום"      
x$words(6)
#> [1] "חשלו"   "לרטי"   "למטכין" "ולחת"   "צורק"   "מנכם"  
x$sentence()
#> [1] "קוויז אדיפיסינג חשלו ישבעס קוויז."
x$paragraph()
#> [1] "אפאח סתשם שבצק אלמנקום קראס קוויז מגמש."
x$paragraphs(3)
#> [1] "שנרא מנורך מורגם בעליק."                                                    
#> [2] "נשואי וסתעד הכייר וסטיבולום להאמית. סיט רוגצה סחטיר קרהשק. אאוגו סטום לרטי."
#> [3] "אמט קוויס צוט ותלברו וואל שנרא אפאח למטכין. למרקוח נולום ברומץ שהכים."      
cat(x$paragraphs(6), sep = "\n")
#> ליאמום הועניב תצטנפל קלאצי ליבם גק בריקנה. לכנו ניסי רוגצה להאמית סטום לכנוץ איאקוליס. מגמש למטכין לתיג יהול מגמש לכנו.
#> בלובק כאנה שמחויט קוויס איף סאפיאן. קונדימנטום עמחליף וק. ולחת סוברט אאוגו ומרגשח. וק מונחף להאמית.
#> תרבנך בלינדו ניבאה הבקיץ כאנה איפסום בגורמי התידם. אמט תצטנפל אאוגו שהכים סת אגת.
#> ותלברו וחאית היושבב להאמית. לפתיעם אמט דולור דז מונחף. שנרא קרהשק להאמית השמה.
#> דולור נון מוסן. הועניב ליץ כאנה קונסקטורר בעריר ישבעס. לורם קלאצי לפתיעם דולור קורוס סטום.
#> דול קרהשק לתיג וחאית. לפתיעם קרהשק הבקיץ בעליק להאמית.
x$text(19)
#> [1] "דולור נובש."
x <- LoremProvider_he_IL$new(word_connector = " --- ")
x$paragraph(4)
#> [1] "סוברט --- לפתיעם --- ארווס --- שנרא --- סכעיט --- זותה --- מנק. --- אדיפיסינג --- ניסי --- ותלברו --- בלובק --- סוליסי --- קלאצי --- קוויס. --- קונסקטורר --- כלרשט --- התידם --- למרקל. --- אט --- דומור --- בלינדו --- וסטיבולום --- בגורמי --- לכנו --- מנכם."