Lorem provider

Methods for Lorem Ipsum generation. Lorem Ipsum is a placeholder text commonly used to demonstrate the visual form of a document or a typeface without relying on meaningful content.

See also

Super classes

charlatan::BareProvider -> charlatan::BaseProvider -> charlatan::LoremProvider -> LoremProvider_zh_CN

Methods

Public methods

LoremProvider_zh_CN$clone()

Inherited methods

charlatan::BareProvider$bothify()

charlatan::BareProvider$lexify()

charlatan::BareProvider$numerify()

charlatan::BareProvider$random_digit()

charlatan::BareProvider$random_digit_not_zero()

charlatan::BareProvider$random_digit_not_zero_or_empty()

charlatan::BareProvider$random_digit_or_empty()

charlatan::BareProvider$random_element()

charlatan::BareProvider$random_element_prob()

charlatan::BareProvider$random_int()

charlatan::BareProvider$random_letter()

charlatan::BareProvider$randomize_nb_elements()

charlatan::BaseProvider$allowed_locales()

charlatan::BaseProvider$check_locale()

charlatan::BaseProvider$print()

charlatan::LoremProvider$initialize()

charlatan::LoremProvider$paragraph()

charlatan::LoremProvider$paragraphs()

charlatan::LoremProvider$sentence()

charlatan::LoremProvider$sentences()

charlatan::LoremProvider$text()

charlatan::LoremProvider$word()

charlatan::LoremProvider$words()

Method clone()

The objects of this class are cloneable with this method.

Usage

LoremProvider_zh_CN$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- LoremProvider_zh_CN$new()
x$word()
#> [1] "更多"
x$words(3)
#> [1] "会员" "就是" "中心"
x$words(6)
#> [1] "表示" "更新" "有些" "国内" "工程" "解决"
x$sentence()
#> [1] "社区 方面 部门 为了 拥有."
x$paragraph()
#> [1] "一般 名称 一直 方式. 一起 历史 数据 完全 怎么. 如何 不同 现在 电子 这个 最新 类型."
x$paragraphs(3)
#> [1] "显示 我的 注册 学习. 进入 搜索 那些 介绍 的话."                              
#> [2] "对于 操作 联系 已经 文件 社区 为什. 经营 今天 回复 不会 一般 是否 不过 活动."
#> [3] "类型 位置 深圳 目前. 部门 经营 美国 作为 中文 这样."                         
cat(x$paragraphs(6), sep = "\n")
#> 科技 没有 部分 无法. 工作 组织 然后 类别.
#> 登录 经营 但是 一直 直接 经营 运行. 学生 这是 工程 那个 使用 经验. 位置 知道 方面.
#> 大学 这个 其实 业务 音乐. 一定 经验 准备 一定.
#> 这里 不过 学校. 内容 工程 很多 什么 其他 科技 为什 责任. 无法 人民 部分 首页 美国 发现 经营.
#> 进入 内容 工程 生活 学习 能够. 决定 是否 原因 技术 如此 首页 也是 类型.
#> 最大 如此 完全 业务. 行业 行业 论坛 查看 这么 发表 地方.
x$text(19)
#> [1] "结果 方法 下载 参加 很多 数据."
x <- LoremProvider_zh_CN$new(word_connector = " --- ")
x$paragraph(4)
#> [1] "进入 --- 学校 --- 然后 --- 结果. --- 网络 --- 为了 --- 浏览 --- 是一 --- 本站. --- 操作 --- 出来 --- 会员 --- 上海 --- 美国 --- 不会 --- 决定 --- 帮助. --- 威望 --- 地址 --- 投资 --- 精华 --- 社区 --- 社会."