PersonProvider

person names methods

Details

Note that with the male/female versions if the locale doesn't provide a male/female version then we fall back to the generic thing, e.g., if no female first name we give you first name

Super class

charlatan::BaseProvider -> PersonProvider

Public fields

locale

(character) the locale

formats

(character) person name formats

person

(character) person name data

messy

(logical) the messy setting, TRUE or FALSE

Methods

Public methods

PersonProvider$allowed_locales()

PersonProvider$new()

PersonProvider$render()

PersonProvider$first_name()

PersonProvider$first_name_female()

PersonProvider$first_name_male()

PersonProvider$last_name()

PersonProvider$last_name_female()

PersonProvider$last_name_male()

PersonProvider$prefix()

PersonProvider$prefix_female()

PersonProvider$prefix_male()

PersonProvider$suffix()

PersonProvider$suffix_female()

PersonProvider$suffix_male()

PersonProvider$clone()

Inherited methods

charlatan::BaseProvider$bothify()

charlatan::BaseProvider$check_locale()

charlatan::BaseProvider$lexify()

charlatan::BaseProvider$numerify()

charlatan::BaseProvider$random_digit()

charlatan::BaseProvider$random_digit_not_zero()

charlatan::BaseProvider$random_digit_not_zero_or_empty()

charlatan::BaseProvider$random_digit_or_empty()

charlatan::BaseProvider$random_element()

charlatan::BaseProvider$random_element_prob()

charlatan::BaseProvider$random_int()

charlatan::BaseProvider$random_letter()

charlatan::BaseProvider$randomize_nb_elements()

Method allowed_locales()

fetch the allowed locales for this provider

Usage

PersonProvider$allowed_locales()

Method new()

Create a new PersonProvider object

Usage

PersonProvider$new(locale = NULL, messy = FALSE)

Arguments

locale

(character) the locale to use. See $allowed_locales() for locales supported (default: en_US)

messy

(logical) make some messy data. Default: FALSE

Returns

A new PersonProvider object

Method render()

Make a person's name

Usage

PersonProvider$render(fmt = NULL)

Arguments

fmt

(character) a name format, default: NULL

Method first_name()

make a first name

Usage

PersonProvider$first_name()

Method first_name_female()

make a female first name

Usage

PersonProvider$first_name_female()

Method first_name_male()

make a male first name

Usage

PersonProvider$first_name_male()

Method last_name()

make a last name

Usage

PersonProvider$last_name()

Method last_name_female()

make a female last name

Usage

PersonProvider$last_name_female()

Method last_name_male()

make a male last name

Usage

PersonProvider$last_name_male()

Method prefix()

make a name prefix

Usage

PersonProvider$prefix()

Method prefix_female()

make a female name prefix

Usage

PersonProvider$prefix_female()

Method prefix_male()

make a male name prefix

Usage

PersonProvider$prefix_male()

Method suffix()

make a name suffix

Usage

PersonProvider$suffix()

Method suffix_female()

make a female name suffix

Usage

PersonProvider$suffix_female()

Method suffix_male()

make a male name suffix

Usage

PersonProvider$suffix_male()

Method clone()

The objects of this class are cloneable with this method.

Usage

PersonProvider$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

x <- PersonProvider$new()
x$locale

#> [1] "en_US"

x$render()

#> [1] "Dr. Dwan Sanford MD"

x$first_name()

#> [1] "Shanell"

x$first_name_female()

#> [1] "Nyasia"

x$first_name_male()

#> [1] "Fount"

x$last_name()

#> [1] "Labadie"

x$last_name_female()

#> [1] "Dach"

x$last_name_male()

#> [1] "Maggio"

x <- PersonProvider$new(locale = "en_GB")
x$locale

#> [1] "en_GB"

x$render()

#> [1] "Gerald Watson"

x$first_name()

#> [1] "Rosemary"

x$first_name_female()

#> [1] "Hazel"

x$first_name_male()

#> [1] "Daniel"

x$last_name()

#> [1] "Wilkins"

x$last_name_female()

#> [1] "Shah"

x$last_name_male()

#> [1] "Burton"

z <- PersonProvider$new(locale = "fr_FR")
z$locale

#> [1] "fr_FR"

z$render()

#> [1] "Emmanuel Boulay"

z$first_name()

#> [1] "Paul"

z$first_name_female()

#> [1] "Élisabeth"

z$first_name_male()

#> [1] "Thibaut"

z$last_name()

#> [1] "Pérrot"

z$last_name_female()

#> [1] "Guyon"

z$last_name_male()

#> [1] "Arnaud"

z$prefix()

#> [1] "d'"

z <- PersonProvider$new(locale = "de_AT")
z$locale

#> [1] "de_AT"

z$render()

#> [1] "Tobias Binder"

z$first_name()

#> [1] "Ingrid"

z$last_name()

#> [1] "Stadler"

z$prefix()

#> [1] "Mag."

z <- PersonProvider$new(locale = "cs_CZ")
z$locale

#> [1] "cs_CZ"

z$render()

#> [1] "paní Jindřiška Králová CSc."

z$first_name()

#> [1] "Emil"

z$first_name_female()

#> [1] "Soňa"

z$first_name_male()

#> [1] "Zbyněk"

z$last_name()

#> [1] "Čermáková"

z$last_name_female()

#> [1] "Krejčová"

z$last_name_male()

#> [1] "Štěpánek"

z$prefix()

#> [1] "paní"

z <- PersonProvider$new(locale = "es_MX")
z$locale

#> [1] "es_MX"

z$render()

#> [1] "Clemente del Valle"

z$first_name()

#> [1] "Hilda"

z$first_name_female()

#> [1] "Esteban"

z$first_name_male()

#> [1] "Paulina"

z$last_name()

#> [1] "Rendón"

z$prefix()

#> [1] "Ing."

z <- PersonProvider$new(locale = "en_NZ")
z$locale

#> [1] "en_NZ"

z$render()

#> [1] "Sarah Watkins"

z$first_name()

#> [1] "Carlos"

z$first_name_female()

#> [1] "Wendy"

z$first_name_male()

#> [1] "Neville"

z$last_name()

#> [1] "Fleming"

PersonProvider$new(locale = "fr_CH")$render()

#> [1] "Denise Cretton"

PersonProvider$new(locale = "fi_FI")$render()

#> [1] "Laila Ahopalo"

PersonProvider$new(locale = "fa_IR")$render()

#> [1] "ساميار اشرفی"

PersonProvider$new(locale = "es_ES")$render()

#> [1] "Aurora Recio Lluch"

PersonProvider$new(locale = "de_DE")$render()

#> [1] "Gert Säuberlich"

PersonProvider$new(locale = "de_AT")$render()

#> [1] "Dr. Michael Lehner"

PersonProvider$new(locale = "cs_CZ")$render()

#> [1] "Daniel Kříž CSc."

PersonProvider$new(locale = "bg_BG")$render()

#> [1] "Флорика Скринска"

PersonProvider$new(locale = "da_DK")$render()

#> [1] "Sys Madsen"