TaxonomyProvider

Taxonomy provider

Names

Names were taken from Theplantlist. 500 genera names and 500 epithets were chosen at random from the set of 10,000 names in the dataset in the taxize package. Theplantlist is, as it says on the tin, composed of plant names - so these fake names are derived from plant names if that matters to you. These may generate names that match those of real taxa, but may not as well.

Taxonomic authority

Randomly, the taxonomic authority is in parentheses - which represents that the given authority was not the original authority.

Super class

charlatan::BaseProvider -> TaxonomyProvider

Public fields

genera

(character) vector of generic names

epithets

(character) vector of eptithet names

Methods

Public methods

TaxonomyProvider$genus()

TaxonomyProvider$epithet()

TaxonomyProvider$species()

TaxonomyProvider$clone()

Inherited methods

charlatan::BaseProvider$bothify()

charlatan::BaseProvider$check_locale()

charlatan::BaseProvider$lexify()

charlatan::BaseProvider$numerify()

charlatan::BaseProvider$random_digit()

charlatan::BaseProvider$random_digit_not_zero()

charlatan::BaseProvider$random_digit_not_zero_or_empty()

charlatan::BaseProvider$random_digit_or_empty()

charlatan::BaseProvider$random_element()

charlatan::BaseProvider$random_element_prob()

charlatan::BaseProvider$random_int()

charlatan::BaseProvider$random_letter()

charlatan::BaseProvider$randomize_nb_elements()

Method genus()

Get a genus name

Usage

TaxonomyProvider$genus()

Method epithet()

Get an epithet name

Usage

TaxonomyProvider$epithet()

Method species()

Get a binomial name (genus + epithet)

Usage

TaxonomyProvider$species(authority = FALSE, date = FALSE)

Arguments

authority

Include authority. default: FALSE

date

Include authority date. If authority = FALSE, this is ignored. default: FALSE

Method clone()

The objects of this class are cloneable with this method.

Usage

TaxonomyProvider$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

(z <- TaxonomyProvider$new())

#> <TaxonomyProvider>
#>   Inherits from: <BaseProvider>
#>   Public:
#>     bothify: function (text = "## ??") 
#>     check_locale: function (x) 
#>     clone: function (deep = FALSE) 
#>     epithet: function () 
#>     epithets: abbeokutae ackermani adolphii adunca aegilicum afzelii a ...
#>     genera: Acanthophyllum Achillea Achyrocline Acmella Aconitum Acr ...
#>     genus: function () 
#>     lexify: function (text = "????") 
#>     numerify: function (text = "###") 
#>     random_digit: function () 
#>     random_digit_not_zero: function () 
#>     random_digit_not_zero_or_empty: function () 
#>     random_digit_or_empty: function () 
#>     random_element: function (x) 
#>     random_element_prob: function (x) 
#>     random_int: function (min = 0, max = 9999, size = 1) 
#>     random_letter: function () 
#>     randomize_nb_elements: function (number = 10, le = FALSE, ge = FALSE, min = NULL, max = NULL) 
#>     species: function (authority = FALSE, date = FALSE) 

z$genus()

#> [1] "Eriocoryne"

z$epithet()

#> [1] "looseri"

z$species()

#> [1] "Asplenium lilacina"

z$species(authority = TRUE)

#> [1] "Cynometra kovacevii Quigley"

## FIXME - datetimeprovider slow - may be related to unix time problem
# z$species(authority = TRUE, date = TRUE)