Fraudster - catch all client to make all types of fake data

Fraudster - catch all client to make all types of fake data

Usage

fraudster(locale = NULL)

Arguments

locale

(character) the locale to use. options: en_US (default), fr_FR, fr_CH, hr_FR, fa_IR, pl_PL, ru_RU, uk_UA, zh_TW.

Examples

# English - the default locale
(x <- fraudster())
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> <fraudster>
#>   locale: en_US
x$job()
#> [1] "Translator"
x$name()
#> [1] "River Lang"
x$color_name()
#> [1] "LightCoral"
x$safe_color_name()
#> [1] "green"
x$hex_color()
#> [1] "#3266B4"
x$safe_hex_color()
#> [1] "#33ff66"
x$rgb_color()
#> [[1]]
#> [1] 169  21  42
#> 
x$rgb_css_color()
#> [1] "rgb(203, 82, 19)"

# different locales
## French
(y <- fraudster(locale = "fr_FR"))
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning:  LoremProvider does not have locale fr_FR, defaulting to en_US locale.
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> <fraudster>
#>   locale: fr_FR
y$job()
#> [1] "Conseiller agricole"

## Croatian
(z <- fraudster(locale = "hr_HR"))
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning:  LoremProvider does not have locale hr_HR, defaulting to en_US locale.
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> <fraudster>
#>   locale: hr_HR
z$job()
#> [1] "Radio inspektor sigurnosti plovidbe"

## Ukranian
(w <- fraudster(locale = "uk_UA"))
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> <fraudster>
#>   locale: uk_UA
w$job()
#> [1] "Вчитель"
w$color_name()
#> [1] "Яскраво-бірюзовий"

# geospatial
x$lat()
#> [1] 75.65936
x$lon()
#> [1] 165.4035
x$position()
#>            [,1]
#> [1,] -124.87382
#> [2,]  -33.15167

# DOIs (Digital Object Identifier)
x$doi()
#> [1] "10.76493/80510-76248-5038"