Create fake taxonomic names

Create fake taxonomic names

Usage

ch_taxonomic_genus(n = 1, locale = "en_US")

ch_taxonomic_epithet(n = 1, locale = "en_US")

ch_taxonomic_species(n = 1, locale = "en_US")

Arguments

n

(integer) number of things to get, any non-negative integer

locale

Locale for provider

Names

Names were taken from Theplantlist. 500 genera names and 500 epithets were chosen at random from the set of 10,000 names in the dataset in the taxize package. Theplantlist is, as it says on the tin, composed of plant names - so these fake names are derived from plant names if that matters to you. These may generate names that match those of real taxa, but may not as well.

Taxonomic authority

Randomly, the taxonomic authority is in parentheses - which represents that the given authority was not the original authority.

See also

Examples

ch_taxonomic_genus()
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> [1] "Ambrosia"
ch_taxonomic_genus(10)
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#>  [1] "Sesamum"       "Argyrothamnia" "Meteorium"     "Cocconerion"  
#>  [5] "Origanum"      "Codonopsis"    "Elymotrigia"   "Antholoma"    
#>  [9] "Deeringia"     "Chrysopogon"  
# or even ch_taxonomic_genus(500)

ch_taxonomic_epithet()
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> [1] "kingii"
ch_taxonomic_epithet(10)
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#>  [1] "petaloidea"       "serrata"          "propinqua"        "ponsiae"         
#>  [5] "cordatum"         "leptopedicellata" "jackiana"         "outeniquense"    
#>  [9] "dapsilis"         "gerlandianum"    
# or even ch_taxonomic_epithet(500)

ch_taxonomic_species()
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> [1] "Prosaptia cruzensis"
ch_taxonomic_species(10)
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#> Warning: Global setting for messy is TRUE
#>  [1] "Monosalpinx abbeokutae"   "Jamesonia ruddianum"     
#>  [3] "Spinifex obovatum"        "Tylimanthus dahuricum"   
#>  [5] "Herniaria diaeolum"       "Platystele hyoserifolium"
#>  [7] "Rinzia rhomboglossa"      "Cullen pseudohypnoides"  
#>  [9] "Phymaspermum madrensis"   "Waitzia lanata"          
# or even ch_taxonomic_species(500)