InternetProvider

internet methods, e.g., email addresses, domain names

Note that if a locale you set doesn't have a locale specific set of data for PersonProvider or CompanyProvider we fall back to en_US Also note that you

Value

A InternetProvider object with specific functions for internet.

See also

Super classes

charlatan::BareProvider -> charlatan::BaseProvider -> InternetProvider

Methods

Public methods

InternetProvider$new()

InternetProvider$to_ascii()

InternetProvider$email()

InternetProvider$safe_email()

InternetProvider$free_email()

InternetProvider$company_email()

InternetProvider$ascii_email()

InternetProvider$ascii_safe_email()

InternetProvider$ascii_free_email()

InternetProvider$ascii_company_email()

InternetProvider$user_name()

InternetProvider$tld()

InternetProvider$free_email_domain()

InternetProvider$url()

InternetProvider$domain_name()

InternetProvider$domain_word()

InternetProvider$ipv4()

InternetProvider$ipv6()

InternetProvider$mac_address()

InternetProvider$uri_page()

InternetProvider$uri_path()

InternetProvider$uri_extension()

InternetProvider$uri()

InternetProvider$slug()

InternetProvider$image_url()

InternetProvider$clone()

Inherited methods

charlatan::BareProvider$bothify()

charlatan::BareProvider$lexify()

charlatan::BareProvider$numerify()

charlatan::BareProvider$random_digit()

charlatan::BareProvider$random_digit_not_zero()

charlatan::BareProvider$random_digit_not_zero_or_empty()

charlatan::BareProvider$random_digit_or_empty()

charlatan::BareProvider$random_element()

charlatan::BareProvider$random_element_prob()

charlatan::BareProvider$random_int()

charlatan::BareProvider$random_letter()

charlatan::BareProvider$randomize_nb_elements()

charlatan::BaseProvider$allowed_locales()

charlatan::BaseProvider$check_locale()

charlatan::BaseProvider$print()

Method new()

Create a new Provider object

Usage

InternetProvider$new()

Returns

A new object

Method to_ascii()

convert to ascii

Usage

InternetProvider$to_ascii(x)

Arguments

x

the stringn to convert to ascii

Method email()

get an email address

Usage

InternetProvider$email(domain = NULL)

Arguments

domain

(character) a domain name, if not given, a random name is chosen

Method safe_email()

get a safe email address

Usage

InternetProvider$safe_email()

Method free_email()

a free email address

Usage

InternetProvider$free_email()

Method company_email()

company email address

Usage

InternetProvider$company_email()

Method ascii_email()

ascii email address

Usage

InternetProvider$ascii_email()

Method ascii_safe_email()

safe ascii email address

Usage

InternetProvider$ascii_safe_email()

Method ascii_free_email()

an ascii free email address

Usage

InternetProvider$ascii_free_email()

Method ascii_company_email()

ascii company email address

Usage

InternetProvider$ascii_company_email()

Method user_name()

a user name

Usage

InternetProvider$user_name()

Method tld()

a tld

Usage

InternetProvider$tld()

Method free_email_domain()

free email domain name

Usage

InternetProvider$free_email_domain()

Method url()

a url

Usage

InternetProvider$url(schemes = NULL)

Arguments

schemes

(character vector) a url scheme, defaults are http and https

Method domain_name()

Produce an Internet domain name with the specified number of subdomain levels

Usage

InternetProvider$domain_name(levels = 1)

Arguments

levels

(integer) how many levels, must be >1

Method domain_word()

a domain word

Usage

InternetProvider$domain_word()

Method ipv4()

an ipv4 address or network

Usage

InternetProvider$ipv4(network = FALSE)

Arguments

network

(logical) produce a network

Method ipv6()

an ipv6 address or network

Usage

InternetProvider$ipv6(network = FALSE)

Arguments

network

(logical) produce a network

Method mac_address()

a mac address

Usage

InternetProvider$mac_address()

Method uri_page()

a uri page

Usage

InternetProvider$uri_page()

Method uri_path()

a uri path

Usage

InternetProvider$uri_path(deep = NULL)

Arguments

deep

how deep to go, an integer, if not given an integer between 1 and 4 (inclusive) is chosen

Method uri_extension()

a uri extension

Usage

InternetProvider$uri_extension()

Method uri()

a uri

Usage

InternetProvider$uri()

Method slug()

a slug

Usage

InternetProvider$slug(value = NULL)

Arguments

value

(character) a string, if given, returns itself, if not, uses LoremProvider to get a random string. default: NULL

Method image_url()

Returns URL to placeholder image - Example: http://placehold.it/640x480

Usage

InternetProvider$image_url(width = NULL, height = NULL)

Arguments

width

image width, in pixels

height

image height, in pixels

Method clone()

The objects of this class are cloneable with this method.

Usage

InternetProvider$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.