LoremProvider

lorem ipsum methods

Super class

charlatan::BaseProvider -> LoremProvider

Public fields

locale

(character) the locale

Methods

Public methods

LoremProvider$allowed_locales()

LoremProvider$new()

LoremProvider$word()

LoremProvider$words()

LoremProvider$sentence()

LoremProvider$sentences()

LoremProvider$paragraph()

LoremProvider$paragraphs()

LoremProvider$text()

LoremProvider$clone()

Inherited methods

charlatan::BaseProvider$bothify()

charlatan::BaseProvider$check_locale()

charlatan::BaseProvider$lexify()

charlatan::BaseProvider$numerify()

charlatan::BaseProvider$random_digit()

charlatan::BaseProvider$random_digit_not_zero()

charlatan::BaseProvider$random_digit_not_zero_or_empty()

charlatan::BaseProvider$random_digit_or_empty()

charlatan::BaseProvider$random_element()

charlatan::BaseProvider$random_element_prob()

charlatan::BaseProvider$random_int()

charlatan::BaseProvider$random_letter()

charlatan::BaseProvider$randomize_nb_elements()

Method allowed_locales()

fetch the allowed locales for this provider

Usage

LoremProvider$allowed_locales()

Method new()

Create a new LoremProvider object

Usage

LoremProvider$new(
  locale = NULL,
  sentence_punctuation = ".",
  word_connector = " "
)

Arguments

locale

(character) the locale to use. See $allowed_locales() for locales supported (default: en_US)

sentence_punctuation

(character) End of sentence punctuation

word_connector

(character) Default connector between words

Returns

A new LoremProvider object

Method word()

Generate a random word

Usage

LoremProvider$word(ext_words = NULL)

Arguments

ext_words

a character vector of words you would like to have instead of 'Lorem ipsum'

Returns

a single word

Method words()

Generate a character vector of random words

Usage

LoremProvider$words(nb = 3, ext_words = NULL)

Arguments

nb

(integer) how many words to return

ext_words

a character vector of words you would like to have instead of 'Lorem ipsum'

Returns

many words

Method sentence()

Generate a random sentence

Usage

LoremProvider$sentence(
  nb_words = 6,
  variable_nb_words = TRUE,
  ext_words = NULL
)

Arguments

nb_words

(integer) around how many words the sentence should contain

variable_nb_words

set to FALSE if you want exactly nb words returned, otherwise the result may include a number of words of nb +/-40% (with a minimum of 1)

ext_words

a character vector of words you would like to have instead of 'Lorem ipsum'

Returns

a single sentence

Method sentences()

Generate a character vector of random sentences

Usage

LoremProvider$sentences(nb = 3, ext_words = NULL)

Arguments

nb

(integer) how many sentences to return

ext_words

a character vector of words you would like to have instead of 'Lorem ipsum'

Returns

many sentences

Method paragraph()

Generate a single paragraph

Usage

LoremProvider$paragraph(
  nb_sentences = 3,
  variable_nb_sentences = TRUE,
  ext_words = NULL
)

Arguments

nb_sentences

(integer) around how many sentences the paragraph should contain

variable_nb_sentences

set to FALSE if you want exactly nb sentences returned, otherwise the result may include a number of sentences of nb +/-40% (with a minimum of 1)

ext_words

a character vector of words you would like to have instead of 'Lorem ipsum'

Returns

a single paragraph

Method paragraphs()

Generate many paragraphs

Usage

LoremProvider$paragraphs(nb = 3, ext_words = NULL)

Arguments

nb

(integer) how many paragraphs to return

ext_words

a character vector of words you would like to have instead of 'Lorem ipsum'

Returns

many paragraphs

Method text()

Generate a random text string. Depending on the max_nb_chars, returns a string made of words, sentences, or paragraphs.

Usage

LoremProvider$text(max_nb_chars = 200, ext_words = NULL)

Arguments

max_nb_chars

Maximum number of characters the text should contain (minimum 5)

ext_words

a character vector of words you would like to have instead of 'Lorem ipsum'

Returns

character string of words

Method clone()

The objects of this class are cloneable with this method.

Usage

LoremProvider$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

(x <- LoremProvider$new())

#> <LoremProvider>
#>   Inherits from: <BaseProvider>
#>   Public:
#>     allowed_locales: function () 
#>     bothify: function (text = "## ??") 
#>     check_locale: function (x) 
#>     clone: function (deep = FALSE) 
#>     initialize: function (locale = NULL, sentence_punctuation = ".", word_connector = " ") 
#>     lexify: function (text = "????") 
#>     locale: en_US
#>     numerify: function (text = "###") 
#>     paragraph: function (nb_sentences = 3, variable_nb_sentences = TRUE, ext_words = NULL) 
#>     paragraphs: function (nb = 3, ext_words = NULL) 
#>     random_digit: function () 
#>     random_digit_not_zero: function () 
#>     random_digit_not_zero_or_empty: function () 
#>     random_digit_or_empty: function () 
#>     random_element: function (x) 
#>     random_element_prob: function (x) 
#>     random_int: function (min = 0, max = 9999, size = 1) 
#>     random_letter: function () 
#>     randomize_nb_elements: function (number = 10, le = FALSE, ge = FALSE, min = NULL, max = NULL) 
#>     sentence: function (nb_words = 6, variable_nb_words = TRUE, ext_words = NULL) 
#>     sentences: function (nb = 3, ext_words = NULL) 
#>     text: function (max_nb_chars = 200, ext_words = NULL) 
#>     word: function (ext_words = NULL) 
#>     words: function (nb = 3, ext_words = NULL) 
#>   Private:
#>     cap_first: function (x) 
#>     drop_last: function (x) 
#>     locales: en_US ar_AA el_GR he_IL ja_JP la ru_RU zh_CN zh_TW
#>     parse_eval_safe: function (name) 
#>     sentence_punctuation: .
#>     word_connector:  
#>     word_list: a ability able about above accept according account acro ...

x$locale

#> [1] "en_US"

x$word()

#> [1] "executive"

x$words(3)

#> [1] "cold"    "similar" "player" 

x$words(6)

#> [1] "hold"      "happen"    "blood"     "do"        "during"    "condition"

x$sentence()

#> [1] "Various discuss major parent."

x$sentences(3)

#> [1] "Prepare bring every raise group although word."
#> [2] "Full lose daughter scientist book so wait."    
#> [3] "People travel young Mr stuff."                 

x$sentences(6)

#> [1] "Defense address public learn service leave."             
#> [2] "Strategy image discover word stuff."                     
#> [3] "Design eat wrong production make."                       
#> [4] "Beyond second same such true remain summer particularly."
#> [5] "Describe water experience great."                        
#> [6] "Environmental though goal result race phone clear."      

x$paragraph()

#> [1] "Majority then president total home cultural maintain."

x$paragraphs(3)

#> [1] "Population information cut car plan guy let visit. Another country however deep argue."                        
#> [2] "Value ask I company forget appear. Card just cultural majority article stop. Country week do young particular."
#> [3] "Small think news impact young officer tough."                                                                  

x$paragraphs(6)

#> [1] "Like prove people foot others. Activity choice daughter ever dark radio politics. Dark yourself guy audience discussion play day."
#> [2] "Anything production people performance maybe. Establish hand smile fish girl good age maintain."                                  
#> [3] "Environment what president professor sell air wide. Spend skin learn action. Morning former standard trouble."                    
#> [4] "Wide almost now reason control edge. Understand purpose contain happen way network hospital by."                                  
#> [5] "Speech make task collection social owner. Others difference son above."                                                           
#> [6] "Ten image myself choose car. Discuss performance movie school one special. Myself Democrat more music."                           

cat(x$paragraphs(6), sep = "\n")

#> Act where pay with simple part. Watch huge record than significant be occur. List individual room white pull state.
#> Develop conference contain. Store American can.
#> American receive skin great. Arrive compare thing game positive significant easy.
#> Simple do possible. Simply husband us edge teacher. Reality church later paper boy. Himself effort tonight specific little.
#> By door drug my night its. Value sort fine professor government ground possible.
#> Industry maybe sport indicate matter still specific. Yourself as Mrs hold ask main. Off improve money shoulder picture raise risk.

x$text(6)

#> [1] "Local."

x$text(10)

#> [1] "History."

x$text(19)

#> [1] "First grow."

x$text(25)

#> [1] "Both suffer event arm."

x$text(50)

#> [1] "North painting responsibility pick."

x$text(300)

#> [1] "Condition compare tend exist unit in crime. Federal beautiful security speech.\nStreet green man season claim. Myself behavior find both.\nMy support wonder. Region air who.\nSon peace bill allow know draw few."

x$text(2000)

#> [1] "Call our me service image hair environmental. And edge development yet. Figure toward level meeting pay consumer worry.\nHead here thought maintain else. Bed whatever they PM citizen. House current name wear letter human course.\nHistory center shake. Page sense type single fast break show.\nSense data guy true argue thus. Attorney major machine hot.\nOrganization room voice account young. Reality message figure nothing enjoy any current kitchen. Administration manage north coach.\nRespond whom push their guy. Around everything rate go politics drive lay per.\nLikely board policy bill church. Travel bar tree cover argue. Push sort along project team my visit family.\nInvestment program population long center. Plant wish turn attack your alone factor financial. Actually catch result certain. Always build term.\nWorry responsibility fall over. Particular decision campaign treatment big hundred. Perform sense arrive PM as view.\nSouthern sing level. Show analysis western wait then land.\nDespite offer wide something trade local data score. Success later truth.\nSome nearly care do born. Interest culture around when American theory now. Plant poor yard next. Him remember event.\nHere scientist if choice against join. Who section wear easy.\nYoung meeting on interesting brother American safe begin. Either pick letter probably pattern degree official pass. Few son usually left.\nPiece place consumer technology training it coach. True age information party his budget necessary. Star indeed show.\nHow pick economy food. Million road surface.\nSkill near sound mention nation doctor shoulder. Agree social government type interest seek almost. Return turn tough right wonder follow.\nTheir full reach week finish. Present entire follow maybe take. Home possible drug.\nAdd race population case. Understand trouble theory bit. Age center school.\nCover claim lose include data. Its week expert operation people item have. Main film safe ok common form."

# set a different sentence_punctuation or word_connector
(x <- LoremProvider$new(sentence_punctuation = ";"))

#> <LoremProvider>
#>   Inherits from: <BaseProvider>
#>   Public:
#>     allowed_locales: function () 
#>     bothify: function (text = "## ??") 
#>     check_locale: function (x) 
#>     clone: function (deep = FALSE) 
#>     initialize: function (locale = NULL, sentence_punctuation = ".", word_connector = " ") 
#>     lexify: function (text = "????") 
#>     locale: en_US
#>     numerify: function (text = "###") 
#>     paragraph: function (nb_sentences = 3, variable_nb_sentences = TRUE, ext_words = NULL) 
#>     paragraphs: function (nb = 3, ext_words = NULL) 
#>     random_digit: function () 
#>     random_digit_not_zero: function () 
#>     random_digit_not_zero_or_empty: function () 
#>     random_digit_or_empty: function () 
#>     random_element: function (x) 
#>     random_element_prob: function (x) 
#>     random_int: function (min = 0, max = 9999, size = 1) 
#>     random_letter: function () 
#>     randomize_nb_elements: function (number = 10, le = FALSE, ge = FALSE, min = NULL, max = NULL) 
#>     sentence: function (nb_words = 6, variable_nb_words = TRUE, ext_words = NULL) 
#>     sentences: function (nb = 3, ext_words = NULL) 
#>     text: function (max_nb_chars = 200, ext_words = NULL) 
#>     word: function (ext_words = NULL) 
#>     words: function (nb = 3, ext_words = NULL) 
#>   Private:
#>     cap_first: function (x) 
#>     drop_last: function (x) 
#>     locales: en_US ar_AA el_GR he_IL ja_JP la ru_RU zh_CN zh_TW
#>     parse_eval_safe: function (name) 
#>     sentence_punctuation: ;
#>     word_connector:  
#>     word_list: a ability able about above accept according account acro ...

x$paragraph(4)

#> [1] "Consumer sort call tree summer financial material scientist; Realize security million source;"

(x <- LoremProvider$new(word_connector = " --- "))

#> <LoremProvider>
#>   Inherits from: <BaseProvider>
#>   Public:
#>     allowed_locales: function () 
#>     bothify: function (text = "## ??") 
#>     check_locale: function (x) 
#>     clone: function (deep = FALSE) 
#>     initialize: function (locale = NULL, sentence_punctuation = ".", word_connector = " ") 
#>     lexify: function (text = "????") 
#>     locale: en_US
#>     numerify: function (text = "###") 
#>     paragraph: function (nb_sentences = 3, variable_nb_sentences = TRUE, ext_words = NULL) 
#>     paragraphs: function (nb = 3, ext_words = NULL) 
#>     random_digit: function () 
#>     random_digit_not_zero: function () 
#>     random_digit_not_zero_or_empty: function () 
#>     random_digit_or_empty: function () 
#>     random_element: function (x) 
#>     random_element_prob: function (x) 
#>     random_int: function (min = 0, max = 9999, size = 1) 
#>     random_letter: function () 
#>     randomize_nb_elements: function (number = 10, le = FALSE, ge = FALSE, min = NULL, max = NULL) 
#>     sentence: function (nb_words = 6, variable_nb_words = TRUE, ext_words = NULL) 
#>     sentences: function (nb = 3, ext_words = NULL) 
#>     text: function (max_nb_chars = 200, ext_words = NULL) 
#>     word: function (ext_words = NULL) 
#>     words: function (nb = 3, ext_words = NULL) 
#>   Private:
#>     cap_first: function (x) 
#>     drop_last: function (x) 
#>     locales: en_US ar_AA el_GR he_IL ja_JP la ru_RU zh_CN zh_TW
#>     parse_eval_safe: function (name) 
#>     sentence_punctuation: .
#>     word_connector:  --- 
#>     word_list: a ability able about above accept according account acro ...

x$paragraph(4)

#> [1] "Own --- question --- way --- away --- guy --- computer. --- Member --- catch --- money --- leg --- rest. --- Strategy --- pull --- car --- difference --- dream --- four --- manage. --- Agree --- former --- authority --- piece --- represent."

# different locales
LoremProvider$new(locale = "ar_AA")$word()

#> [1] "أملاً"

LoremProvider$new(locale = "el_GR")$word()

#> [1] "εντολές"

LoremProvider$new(locale = "he_IL")$word()

#> [1] "להאמית"

LoremProvider$new(locale = "ja_JP")$word()

#> [1] "販売"

LoremProvider$new(locale = "zh_TW")$word()

#> [1] "廣告"