charlatan

Make fake data, supporting addresses, person names, dates, times, colors, coordinates, currencies, digital object identifiers (DOIs), jobs, phone numbers, DNA sequences, doubles and integers from distributions and within a range.

Package API

ch_generate(): generate a data.frame with fake data

fraudster(): single interface to all fake data methods

High level interfaces: There are high level functions prefixed with ch_ that wrap low level interfaces, and are meant to be easier to use and provide easy way to make many instances of a thing.

Low level interfaces: All of these are R6 objects that a user can initialize and then call methods on the them.

Author

Scott Chamberlain myrmecocystus+r@gmail.com

Kyle Voytovich

Martin Pedersen

Examples

# generate individual types of data
ch_name()

#> [1] "Benji Mayer-Stoltenberg"

ch_phone_number()

#> [1] "+53(4)3274934299"

ch_job()

#> [1] "Designer, television/film set"

# generate a data.frame
ch_generate()

#> # A tibble: 10 x 3
#>    name                     job                              phone_number       
#>    <chr>                    <chr>                            <chr>              
#>  1 Shaquille Parisian-Ortiz Broadcast presenter              1-106-142-5196x306 
#>  2 Jasmyne Mitchell         Therapist, occupational          1-763-465-1011x470…
#>  3 Aaden Hills              Environmental health practition… 1-347-663-6406x725…
#>  4 Heath Fritsch            Research officer, trade union    +69(1)7616102143   
#>  5 Miss Elaine VonRueden D… Geologist, wellsite              632.890.6277x09877 
#>  6 Jadiel Effertz           Community arts worker            1-931-486-0670     
#>  7 Jaycob Ebert             Psychiatric nurse                980-351-4686       
#>  8 Ms. Laureen Bogisich MD  Counsellor                       1-441-196-4417x680…
#>  9 Pate Champlin-Roberts    Engineer, automotive             385.706.8477       
#> 10 Herminia Jast DDS        Scientist, physiological         (802)479-7031x41499

# one interface to all data types - generate the class first
#  reports the locale to be used, can change optionally
(x <- fraudster())

#> <fraudster>
#>   locale: en_US

x$job()

#> [1] "Scientist, biomedical"

x$name()

#> [1] "Arlie Howell"

x$color_name()

#> [1] "Moccasin"

x$hex_color()

#> [1] "#9D1225"

# low level interfaces to "data providers"
# these are exported by hidden from package man page
# as most users will likely not interact with these
x <- ColorProvider$new()
x$color_name()

#> [1] "OrangeRed"

x$hex_color()

#> [1] "#91EAD0"